Fix thread safety issue with Entity
